Improve handling of docker_work.
Previosuly if the mount failed we would not retry to mount the folder.
Leaving users with a docker_work folder inside the container that did not work
and would generate permission errors.

Now we simply try to mount the folder each time we login, and supress the error
if it is allready mounted.